Reverted move of provider and component
