+++ common/v2/features/Toasts/components/Toast.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
 
-import { Typography } from '.';
+import Typography from 'v2/components/Typography';
 import { BREAK_POINTS, COLORS } from 'v2/theme';
 
 import { ToastConfig, ToastType } from 'v2/types';
