+++ common/v2/features/index.ts
@@ -28,3 +28,4 @@ export { SignAndVerifyMessage } from './SignAndVerifyMessage';
 export { NotificationsProvider } from './NotificationsPanel';
 export { BroadcastTransactionFlow } from './BroadcastTransaction';
 export { SwapAssetsFlow } from './SwapAssets';
+export { ToastProvider } from './Toasts';
